requirements.txt
MetaTrader5==5.0.4874
numpy==1.23.5
pandas==2.0.3
matplotlib==3.7.3
scikit-learn==1.3.2
PyYAML==6.0.2
tqdm==4.67.1
bayesian-optimization==1.4.3
tenacity==9.0.0
singleton-decorator==1.0.0
tensorflow==2.12.0
tk==8.6.13